Refactor ListItemForm and ListingDetails components
